Implement optional user logout with systemd and libproc
When we pass "Y" to the logout_user column,
we always first try to find the systemd library
and open it with dlopen, then we call TerminateUser.
If it fails, we use the fallback method which scans the processes
with uid equal to the user to be locked.

In any case a last scan of the processes is done to see if there's
still some process running, in that case the insert returns with an
error.